Reorganized class files and changed character position data members from int to float.
The main reorganization is the removal of the Grunt class, with the intentions of having all enemy types classified by an int value rather than their own class. There was no real need to have each enemy type have their own class and only served to complicate certain functions.

The switch from int to float for the position values of all characters was so that when NPCs move towards a moving target (i.e. the player), they would move in a straight line towards thier target rather than stumbling off the path as they did.

+++ Headers/map.h
@@ -33,7 +33,7 @@ public:
 	// Renders the map.
 	void render(SDL_Renderer* renderer);
 	// Set the x and y rendering position of the map such that the player's location on the map is at the centre of the renderer.
-	void setCentrePlayer(int playerX, int playerY, ScreenDimensions screen);
+	void setCentrePlayer(float playerX, float playerY, ScreenDimensions screen);
 	// Getters
 	SDL_Texture* getFloor();
 	int getX();
